refactor: decouple
