+++ pkg/registry/apis/dashboard/legacy/queries.go
@@ -28,8 +28,6 @@ func mustTemplate(filename string) *template.Template {
 var (
 	sqlQueryDashboards = mustTemplate("query_dashboards.sql")
 	sqlQueryPanels     = mustTemplate("query_panels.sql")
-	sqlQueryPlaylists  = mustTemplate("query_playlists.sql")
-	sqlQueryShortURLs  = mustTemplate("query_shorturls.sql")
 )
 
 type sqlQuery struct {
@@ -83,55 +81,5 @@ func newLibraryQueryReq(sql *legacysql.LegacyDatabaseHelper, query *LibraryPanel
 
 		LibraryElementTable: sql.Table("library_element"),
 		UserTable:           sql.Table("user"),
-	}
-}
-
-type PlaylistQuery struct {
-	OrgID int64
-}
-
-type sqlPlaylistQuery struct {
-	sqltemplate.SQLTemplate
-	Query *PlaylistQuery
-
-	PlaylistTable     string
-	PlaylistItemTable string
-}
-
-func (r sqlPlaylistQuery) Validate() error {
-	return nil
-}
-
-func newPlaylistQueryReq(sql *legacysql.LegacyDatabaseHelper, query *PlaylistQuery) sqlPlaylistQuery {
-	return sqlPlaylistQuery{
-		SQLTemplate: sqltemplate.New(sql.DialectForDriver()),
-		Query:       query,
-
-		PlaylistTable:     sql.Table("playlist"),
-		PlaylistItemTable: sql.Table("playlist_item"),
-	}
-}
-
-type ShortURLQuery struct {
-	OrgID int64
-}
-
-type sqlShortURLQuery struct {
-	sqltemplate.SQLTemplate
-	Query *ShortURLQuery
-
-	ShortURLTable string
-}
-
-func (r sqlShortURLQuery) Validate() error {
-	return nil
-}
-
-func newShortURLQueryReq(sql *legacysql.LegacyDatabaseHelper, query *ShortURLQuery) sqlShortURLQuery {
-	return sqlShortURLQuery{
-		SQLTemplate: sqltemplate.New(sql.DialectForDriver()),
-		Query:       query,
-
-		ShortURLTable: sql.Table("short_url"),
 	}
 }
