+++ pkg/registry/apis/dashboard/legacy/sql_dashboards.go
@@ -23,8 +23,6 @@ import (
 	dashboardV1 "github.com/grafana/grafana/apps/dashboard/pkg/apis/dashboard/v1beta1"
 	"github.com/grafana/grafana/apps/dashboard/pkg/migration/schemaversion"
 	folders "github.com/grafana/grafana/apps/folder/pkg/apis/folder/v1beta1"
-	playlistv1 "github.com/grafana/grafana/apps/playlist/pkg/apis/playlist/v1"
-	shorturlv1beta1 "github.com/grafana/grafana/apps/shorturl/pkg/apis/shorturl/v1beta1"
 	"github.com/grafana/grafana/pkg/apimachinery/apis/common/v0alpha1"
 	"github.com/grafana/grafana/pkg/apimachinery/identity"
 	"github.com/grafana/grafana/pkg/apimachinery/utils"
@@ -42,7 +40,7 @@ import (
 	"github.com/grafana/grafana/pkg/services/search/sort"
 	"github.com/grafana/grafana/pkg/services/sqlstore"
 	"github.com/grafana/grafana/pkg/storage/legacysql"
-	"github.com/grafana/grafana/pkg/storage/unified/migrations/resources"
+	"github.com/grafana/grafana/pkg/storage/unified/migrations"
 	"github.com/grafana/grafana/pkg/storage/unified/resource"
 	"github.com/grafana/grafana/pkg/storage/unified/resourcepb"
 	"github.com/grafana/grafana/pkg/storage/unified/sql/sqltemplate"
@@ -51,9 +49,6 @@ import (
 var (
 	tracer = otel.Tracer("github.com/grafana/grafana/pkg/registry/apis/dashboard/legacy")
 )
-
-// MigrateOptions is an alias for the type defined in the resources package.
-type MigrateOptions = resources.MigrateOptions
 
 type dashboardRow struct {
 	// The numeric version for this dashboard
@@ -89,18 +84,15 @@ type dashboardSqlAccess struct {
 	log         log.Logger
 }
 
-// ProvideResourceMigrationService creates the central ResourceMigrationService for
-// unified storage migrations. It includes the minimal dependencies needed to migrate
-// all resource types (dashboards, folders, library panels, playlists).
-func ProvideResourceMigrationService(
+func ProvideMigrator(
 	sql legacysql.LegacyDatabaseProvider,
 	provisioning provisioning.StubProvisioningService,
 	accessControl accesscontrol.AccessControl,
-) resources.ResourceMigrationService {
-	return newMigratorAccess(sql, provisioning, accessControl)
+) Migrator {
+	return NewMigratorAccess(sql, provisioning, accessControl)
 }
 
-func newMigratorAccess(
+func NewMigratorAccess(
 	sql legacysql.LegacyDatabaseProvider,
 	provisioning provisioning.StubProvisioningService,
 	accessControl accesscontrol.AccessControl,
@@ -207,7 +199,7 @@ func (a *dashboardSqlAccess) getRows(ctx context.Context, helper *legacysql.Lega
 }
 
 // CountResources counts resources without migrating them
-func (a *dashboardSqlAccess) CountResources(ctx context.Context, opts MigrateOptions) (*resourcepb.BulkResponse, error) {
+func (a *dashboardSqlAccess) CountResources(ctx context.Context, opts migrations.MigrateOptions) (*resourcepb.BulkResponse, error) {
 	sql, err := a.sql(ctx)
 	if err != nil {
 		return nil, err
@@ -275,7 +267,7 @@ func (a *dashboardSqlAccess) CountResources(ctx context.Context, opts MigrateOpt
 }
 
 // MigrateDashboards handles the dashboard migration logic
-func (a *dashboardSqlAccess) MigrateDashboards(ctx context.Context, orgId int64, opts MigrateOptions, stream resourcepb.BulkStore_BulkProcessClient) error {
+func (a *dashboardSqlAccess) MigrateDashboards(ctx context.Context, orgId int64, opts migrations.MigrateOptions, stream resourcepb.BulkStore_BulkProcessClient) error {
 	query := &DashboardQuery{
 		OrgID:         orgId,
 		Limit:         100000000,
@@ -367,7 +359,7 @@ func (a *dashboardSqlAccess) MigrateDashboards(ctx context.Context, orgId int64,
 }
 
 // MigrateFolders handles the folder migration logic
-func (a *dashboardSqlAccess) MigrateFolders(ctx context.Context, orgId int64, opts MigrateOptions, stream resourcepb.BulkStore_BulkProcessClient) error {
+func (a *dashboardSqlAccess) MigrateFolders(ctx context.Context, orgId int64, opts migrations.MigrateOptions, stream resourcepb.BulkStore_BulkProcessClient) error {
 	query := &DashboardQuery{
 		OrgID:      orgId,
 		Limit:      100000000,
@@ -450,7 +442,7 @@ func (a *dashboardSqlAccess) MigrateFolders(ctx context.Context, orgId int64, op
 }
 
 // MigrateLibraryPanels handles the library panel migration logic
-func (a *dashboardSqlAccess) MigrateLibraryPanels(ctx context.Context, orgId int64, opts MigrateOptions, stream resourcepb.BulkStore_BulkProcessClient) error {
+func (a *dashboardSqlAccess) MigrateLibraryPanels(ctx context.Context, orgId int64, opts migrations.MigrateOptions, stream resourcepb.BulkStore_BulkProcessClient) error {
 	opts.Progress(-1, "migrating library panels...")
 	panels, err := a.GetLibraryPanels(ctx, LibraryPanelQuery{
 		OrgID: orgId,
@@ -495,137 +487,6 @@ func (a *dashboardSqlAccess) MigrateLibraryPanels(ctx context.Context, orgId int
 		}
 	}
 	opts.Progress(-2, fmt.Sprintf("finished panels... (%d)", len(panels.Items)))
-	return nil
-}
-
-// MigratePlaylists handles the playlist migration logic
-func (a *dashboardSqlAccess) MigratePlaylists(ctx context.Context, orgId int64, opts MigrateOptions, stream resourcepb.BulkStore_BulkProcessClient) error {
-	opts.Progress(-1, "migrating playlists...")
-	rows, err := a.ListPlaylists(ctx, orgId)
-	if rows != nil {
-		defer func() {
-			_ = rows.Close()
-		}()
-	}
-	if err != nil {
-		return err
-	}
-
-	// Group playlist items by playlist ID while preserving order
-	type playlistData struct {
-		id        int64
-		uid       string
-		name      string
-		interval  string
-		items     []playlistv1.PlaylistItem
-		createdAt int64
-		updatedAt int64
-	}
-
-	playlistIndex := make(map[int64]int) // maps playlist ID to index in playlists slice
-	playlists := []*playlistData{}
-	var currentID int64
-	var orgID int64
-	var uid, name, interval string
-	var createdAt, updatedAt int64
-	var itemType, itemValue sql.NullString
-
-	count := 0
-	for rows.Next() {
-		err = rows.Scan(&currentID, &orgID, &uid, &name, &interval, &createdAt, &updatedAt, &itemType, &itemValue)
-		if err != nil {
-			return err
-		}
-
-		// Get or create playlist entry
-		idx, exists := playlistIndex[currentID]
-		var pl *playlistData
-		if !exists {
-			pl = &playlistData{
-				id:        currentID,
-				uid:       uid,
-				name:      name,
-				interval:  interval,
-				items:     []playlistv1.PlaylistItem{},
-				createdAt: createdAt,
-				updatedAt: updatedAt,
-			}
-			playlistIndex[currentID] = len(playlists)
-			playlists = append(playlists, pl)
-		} else {
-			pl = playlists[idx]
-		}
-
-		// Add item if it exists (LEFT JOIN can return NULL for playlists without items)
-		if itemType.Valid && itemValue.Valid {
-			pl.items = append(pl.items, playlistv1.PlaylistItem{
-				Type:  playlistv1.PlaylistPlaylistItemType(itemType.String),
-				Value: itemValue.String,
-			})
-		}
-	}
-
-	if err = rows.Err(); err != nil {
-		return err
-	}
-
-	// Convert to K8s objects and send to stream (order is preserved)
-	for _, pl := range playlists {
-		playlist := &playlistv1.Playlist{
-			TypeMeta: metav1.TypeMeta{
-				APIVersion: playlistv1.GroupVersion.String(),
-				Kind:       "Playlist",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:              pl.uid,
-				Namespace:         opts.Namespace,
-				CreationTimestamp: metav1.NewTime(time.UnixMilli(pl.createdAt)),
-			},
-			Spec: playlistv1.PlaylistSpec{
-				Title:    pl.name,
-				Interval: pl.interval,
-				Items:    pl.items,
-			},
-		}
-
-		// Set updated timestamp if different from created
-		if pl.updatedAt != pl.createdAt {
-			meta, err := utils.MetaAccessor(playlist)
-			if err != nil {
-				return err
-			}
-			updatedTime := time.UnixMilli(pl.updatedAt)
-			meta.SetUpdatedTimestamp(&updatedTime)
-		}
-
-		body, err := json.Marshal(playlist)
-		if err != nil {
-			return err
-		}
-
-		req := &resourcepb.BulkRequest{
-			Key: &resourcepb.ResourceKey{
-				Namespace: opts.Namespace,
-				Group:     "playlist.grafana.app",
-				Resource:  "playlists",
-				Name:      pl.uid,
-			},
-			Value:  body,
-			Action: resourcepb.BulkRequest_ADDED,
-		}
-
-		opts.Progress(count, fmt.Sprintf("%s (%d)", pl.name, len(req.Value)))
-		count++
-
-		err = stream.Send(req)
-		if err != nil {
-			if errors.Is(err, io.EOF) {
-				err = nil
-			}
-			return err
-		}
-	}
-	opts.Progress(-2, fmt.Sprintf("finished playlists... (%d)", len(playlists)))
 	return nil
 }
 
@@ -1322,141 +1183,4 @@ func parseLibraryPanelRow(p panel) (dashboardV0.LibraryPanel, error) {
 
 func (b *dashboardSqlAccess) RebuildIndexes(ctx context.Context, req *resourcepb.RebuildIndexesRequest) (*resourcepb.RebuildIndexesResponse, error) {
 	return nil, fmt.Errorf("not implemented")
-}
-
-func (a *dashboardSqlAccess) ListPlaylists(ctx context.Context, orgID int64) (*sql.Rows, error) {
-	ctx, span := tracer.Start(ctx, "legacy.dashboardSqlAccess.ListPlaylists")
-	defer span.End()
-
-	helper, err := a.sql(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	req := newPlaylistQueryReq(helper, &PlaylistQuery{
-		OrgID: orgID,
-	})
-
-	rawQuery, err := sqltemplate.Execute(sqlQueryPlaylists, req)
-	if err != nil {
-		return nil, fmt.Errorf("execute template %q: %w", sqlQueryPlaylists.Name(), err)
-	}
-
-	rows, err := a.executeQuery(ctx, helper, rawQuery, req.GetArgs()...)
-	if err != nil && rows != nil {
-		_ = rows.Close()
-		return nil, err
-	}
-	return rows, err
-}
-
-func (a *dashboardSqlAccess) ListShortURLs(ctx context.Context, orgID int64) (*sql.Rows, error) {
-	ctx, span := tracer.Start(ctx, "legacy.dashboardSqlAccess.ListShortURLs")
-	defer span.End()
-
-	helper, err := a.sql(ctx)
-	if err != nil {
-		return nil, err
-	}
-
-	req := newShortURLQueryReq(helper, &ShortURLQuery{
-		OrgID: orgID,
-	})
-
-	rawQuery, err := sqltemplate.Execute(sqlQueryShortURLs, req)
-	if err != nil {
-		return nil, fmt.Errorf("execute template %q: %w", sqlQueryShortURLs.Name(), err)
-	}
-
-	rows, err := a.executeQuery(ctx, helper, rawQuery, req.GetArgs()...)
-	if err != nil && rows != nil {
-		_ = rows.Close()
-		return nil, err
-	}
-	return rows, err
-}
-
-// MigrateShortURLs handles the short URL migration logic
-func (a *dashboardSqlAccess) MigrateShortURLs(ctx context.Context, orgId int64, opts MigrateOptions, stream resourcepb.BulkStore_BulkProcessClient) error {
-	opts.Progress(-1, "migrating short URLs...")
-	rows, err := a.ListShortURLs(ctx, orgId)
-	if rows != nil {
-		defer func() {
-			_ = rows.Close()
-		}()
-	}
-	if err != nil {
-		return err
-	}
-
-	var id int64
-	var orgID int64
-	var uid, path string
-	var createdBy int64
-	var createdAt int64
-	var lastSeenAt int64
-
-	count := 0
-	for rows.Next() {
-		err = rows.Scan(&id, &orgID, &uid, &path, &createdBy, &createdAt, &lastSeenAt)
-		if err != nil {
-			return err
-		}
-
-		shortURL := &shorturlv1beta1.ShortURL{
-			TypeMeta: metav1.TypeMeta{
-				APIVersion: shorturlv1beta1.GroupVersion.String(),
-				Kind:       "ShortURL",
-			},
-			ObjectMeta: metav1.ObjectMeta{
-				Name:              uid,
-				Namespace:         opts.Namespace,
-				CreationTimestamp: metav1.NewTime(time.Unix(createdAt, 0)),
-			},
-			Spec: shorturlv1beta1.ShortURLSpec{
-				Path: path,
-			},
-			Status: shorturlv1beta1.ShortURLStatus{
-				LastSeenAt: lastSeenAt,
-			},
-		}
-
-		if createdBy > 0 {
-			shortURL.SetCreatedBy(claims.NewTypeID(claims.TypeUser, strconv.FormatInt(createdBy, 10)))
-		}
-
-		body, err := json.Marshal(shortURL)
-		if err != nil {
-			return err
-		}
-
-		req := &resourcepb.BulkRequest{
-			Key: &resourcepb.ResourceKey{
-				Namespace: opts.Namespace,
-				Group:     shorturlv1beta1.APIGroup,
-				Resource:  "shorturls",
-				Name:      uid,
-			},
-			Value:  body,
-			Action: resourcepb.BulkRequest_ADDED,
-		}
-
-		opts.Progress(count, fmt.Sprintf("%s (%d)", uid, len(req.Value)))
-		count++
-
-		err = stream.Send(req)
-		if err != nil {
-			if errors.Is(err, io.EOF) {
-				err = nil
-			}
-			return err
-		}
-	}
-
-	if err = rows.Err(); err != nil {
-		return err
-	}
-
-	opts.Progress(-2, fmt.Sprintf("finished short URLs... (%d)", count))
-	return nil
 }
