+++ pkg/registry/apis/dashboard/migration_registrar.go
@@ -3,17 +3,17 @@ package dashboard
 import (
 	v1beta1 "github.com/grafana/grafana/apps/dashboard/pkg/apis/dashboard/v1beta1"
 	folders "github.com/grafana/grafana/apps/folder/pkg/apis/folder/v1beta1"
+	"github.com/grafana/grafana/pkg/registry/apis/dashboard/migrator"
 	"github.com/grafana/grafana/pkg/storage/unified/migrations"
-	"github.com/grafana/grafana/pkg/storage/unified/migrations/resources"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 )
 
 /*
 FoldersDashboardsMigration returns the migration definition for folders and dashboards.
-This is owned by the dashboard team and uses the ResourceMigrationService
+This is owned by the dashboard team and uses the DashboardMigrator
 to stream folder and dashboard resources from legacy SQL storage.
 */
-func FoldersDashboardsMigration(accessor resources.ResourceMigrationService) migrations.MigrationDefinition {
+func FoldersDashboardsMigration(migrator migrator.FoldersDashboardsMigrator) migrations.MigrationDefinition {
 	folderGR := schema.GroupResource{Group: folders.GROUP, Resource: folders.RESOURCE}
 	dashboardGR := schema.GroupResource{Group: v1beta1.GROUP, Resource: v1beta1.DASHBOARD_RESOURCE}
 
@@ -25,8 +25,8 @@ func FoldersDashboardsMigration(accessor resources.ResourceMigrationService) mig
 			{GroupResource: dashboardGR, LockTable: "dashboard"},
 		},
 		Migrators: map[schema.GroupResource]migrations.MigratorFunc{
-			folderGR:    accessor.MigrateFolders,
-			dashboardGR: accessor.MigrateDashboards,
+			folderGR:    migrator.MigrateFolders,
+			dashboardGR: migrator.MigrateDashboards,
 		},
 		Validators: []migrations.ValidatorFactory{
 			migrations.CountValidation(folderGR, "dashboard", "org_id = ? AND is_folder = true AND deleted IS NULL"),
