+++ pkg/registry/apps/playlist/migrator/testdata/mysql--query_playlists-list.sql
