+++ pkg/registry/apps/playlist/migrator/testdata/sqlite--query_playlists-list.sql
